refactor: remove `is_numpy_buffer` et al. to `ak.nplike`
