+++ src/awkward/_v2/index.py
@@ -160,7 +160,8 @@ class Index:
 
     def __getitem__(self, where):
         out = self._data[where]
-        from awkward._v2._util import is_cupy_buffer, is_jax_buffer
+        from awkward.nplike import is_jax_buffer
+        from awkward.nplike import is_cupy_buffer
 
         if hasattr(out, "shape") and len(out.shape) != 0:
             return Index(out, metadata=self.metadata, nplike=self.nplike)
